refactor(test): merge cli_behavioral_test.rs into behavioral/cli/ structure
Move unique tests from the monolithic cli_behavioral_test.rs (1324 lines)
into the existing behavioral/cli/ module structure:

- Create batch_test.rs (4 tests: stdin/file batch processing, empty stdin/file)
- Create sitemap_test.rs (1 test: --sitemap-url explicit sitemap discovery)
- Add selector tests to single_page_test.rs (h3, table, no-selector)
- Add PDF binary download test to download_test.rs
- Add output-dir flag test to error_path_test.rs
- Remove cli_behavioral_test.rs and its [[test]] entry
- Delete 6 obsolete cli_behavioral__ snapshots, accept 3 new behavioral__ snapshots

Duplicate tests (exit codes, dry-run, crawl, obsidian, download-images,
download-documents) are NOT migrated — they already exist in the target
files with equivalent or more comprehensive coverage.
